Move log level to about page
