+++ Flow.Launcher/SettingPages/ViewModels/SettingsPaneGeneralViewModel.cs
@@ -1,13 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Windows.Forms;
 using CommunityToolkit.Mvvm.Input;
 using Flow.Launcher.Core;
 using Flow.Launcher.Core.Configuration;
 using Flow.Launcher.Core.Resource;
 using Flow.Launcher.Helper;
-using Flow.Launcher.Infrastructure.Logger;
 using Flow.Launcher.Infrastructure.UserSettings;
 using Flow.Launcher.Plugin;
 using Flow.Launcher.Plugin.SharedModels;
@@ -32,7 +30,6 @@ public partial class SettingsPaneGeneralViewModel : BaseModel
     public class SearchWindowAlignData : DropdownDataGeneric<SearchWindowAligns> { }
     public class SearchPrecisionData : DropdownDataGeneric<SearchPrecisionScore> { }
     public class LastQueryModeData : DropdownDataGeneric<LastQueryMode> { }
-    public class LogLevelData : DropdownDataGeneric<LOGLEVEL> { }
 
     public bool StartFlowLauncherOnSystemStartup
     {
@@ -145,16 +142,12 @@ public partial class SettingsPaneGeneralViewModel : BaseModel
     public List<LastQueryModeData> LastQueryModes { get; } =
         DropdownDataGeneric<LastQueryMode>.GetValues<LastQueryModeData>("LastQuery");
 
-    public List<LogLevelData> LogLevels { get; } =
-        DropdownDataGeneric<LOGLEVEL>.GetValues<LogLevelData>("LogLevel");
-
     private void UpdateEnumDropdownLocalizations()
     {
         DropdownDataGeneric<SearchWindowScreens>.UpdateLabels(SearchWindowScreens);
         DropdownDataGeneric<SearchWindowAligns>.UpdateLabels(SearchWindowAligns);
         DropdownDataGeneric<SearchPrecisionScore>.UpdateLabels(SearchPrecisionScores);
         DropdownDataGeneric<LastQueryMode>.UpdateLabels(LastQueryModes);
-        DropdownDataGeneric<LOGLEVEL>.UpdateLabels(LogLevels);
     }
 
     public string Language
@@ -218,22 +211,6 @@ public partial class SettingsPaneGeneralViewModel : BaseModel
             if (value)
             {
                 UpdateApp();
-            }
-        }
-    }
-
-    public LOGLEVEL LogLevel
-    {
-        get => Settings.LogLevel;
-        set
-        {
-            if (Settings.LogLevel != value)
-            {
-                Settings.LogLevel = value;
-
-                Log.SetLogLevel(value);
-
-                UpdateEnumDropdownLocalizations();
             }
         }
     }
