+++ Flow.Launcher/SettingPages/Views/SettingsPaneAbout.xaml
@@ -123,6 +123,14 @@
                 </StackPanel>
             </cc:Card>
 
+            <cc:Card Title="{DynamicResource logLevel}">
+                <ComboBox
+                    DisplayMemberPath="Display"
+                    ItemsSource="{Binding LogLevels}"
+                    SelectedValue="{Binding LogLevel}"
+                    SelectedValuePath="Value" />
+            </cc:Card>
+
             <TextBlock
                 Margin="14 20 0 0"
                 HorizontalAlignment="Center"
